Server fix
Add scheduling to db update to run once daily
Retrying to save on thread collisions now working
