+++ backend/src/main/java/com/movieknights/server/MovieknightsApplication.java
@@ -7,6 +7,6 @@ import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfi
 @SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 public class MovieknightsApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(MovieknightsApplication.class, args);
+				SpringApplication.run(MovieknightsApplication.class, args);
 	}
 }
