+++ backend/src/main/java/com/movieknights/server/setup/DownloadMovieTask.java
@@ -2,9 +2,8 @@ package com.movieknights.server.setup;
 
 import com.movieknights.server.entities.Movie;
 import com.movieknights.server.services.MovieService;
-import com.movieknights.server.utils.TextUtil;
 
-public class DownloadMovieTask implements Runnable {
+  public class DownloadMovieTask implements Runnable {
   private long movieId;
 
   private MovieService movieService;
@@ -15,10 +14,7 @@ public class DownloadMovieTask implements Runnable {
   }
 
   @Override
-  public void run() {
+    public void run() {
     Movie m = movieService.getMovieById(movieId);
-    if (m != null) {
-      System.out.printf("Movie '%s' (id: %s) is now in the database...\n", TextUtil.pimpString(m.getTitle(), TextUtil.LEVEL_INFO), m.getMovieId());
-    }
   }
 }
